Introduction
The	following	document will	explain and guide you through the use of a
shared document created in Google Drive that is tied to	Canvas. Canvas
data can be added to a Google doc using the Canvas API. To	access your
course you will need to create what is called a Canvas Token. This
token is used to authorize access to your course data. The Google
spreadsheet is scripted to search and enter information about the
enrollment,	activity, and assignments from a Canvas course.